requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lxml==4.4.2
mysql-connector==2.2.9
PyYAML==5.3
requests==2.22.0
urllib3==1.25.7
